+++ packages/toolkit/src/db/migrations/0004_next_ikaris.sql
@@ -1,0 +1,2 @@
+DROP TABLE "jobs" CASCADE;--> statement-breakpoint
+DROP TABLE "pattern_jobs" CASCADE;